Report RFID Based Access Control System.
RFID Based Access Control System. The objective of this project is to style a “RFID based access control system”. The sole aim of this project is to avoid the entry of any unwanted/burglar person to some restricted area along side the study of RFID and its use in this type of application.